pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "mini-rag"
version = "1.0.0"
description = "A local mini-RAG system without API keys"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "pdfplumber>=0.11.0",
    "sentence-transformers>=2.3.1",
    "faiss-cpu>=1.7.4",
    "numpy>=1.24.3",
    "scikit-learn>=1.3.2",
    "tqdm>=4.66.1",
]

[project.optional-dependencies]
ui = ["streamlit>=1.29.0"]
dev = [
    "pytest>=7.4.3",
    "pytest-cov>=4.1.0",
    "black>=23.12.1",
    "mypy>=1.7.1",
    "flake8>=6.1.0",
]

[tool.black]
line-length = 100
target-version = ['py310']

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]